Adding unit test and finishing the whole flow

+++ src/main/java/com/rafael/hexagonal/infrastructure/adapter/output/SpringDataTaskRepository.java
@@ -2,5 +2,5 @@ package com.rafael.hexagonal.infrastructure.adapter.output;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpringDataTaskRepository extends JpaRepository<TaskEntity, Integer> {
+public interface SpringDataTaskRepository extends JpaRepository<TaskEntity, Long> {
 }
